refactor: retract internal modules from the public API (#67)
The library exposed ~158 pub items across net/protocol/reporter/stream/
utils/etc. that are implementation detail — every one a semver commitment.
Flip the internal modules to `pub(crate)` so the public surface is just the
intended API; the few genuinely-public types are re-exported at the crate root.

- lib.rs: `pub mod` -> `pub(crate) mod` for auth, cpu, net, protocol,
  reporter, stream, tcp_info, units, utils. `json_report` stays
  `#[doc(hidden)] pub` (the iperf3-schema result model). Re-export
  protocol::{TestResultsJson, StreamResultJson} (the `run()` return model)
  alongside the existing TransportProtocol / set_cpu_affinity.
- TestConfig::from_params -> pub(crate) (it takes the internal TestParams).
- Migrate 36 internal-reaching tests from tests/integration.rs in-crate into
  the relevant module's #[cfg(test)] block (an external test crate can no
  longer reach `riperf3::protocol::*` etc.). Public-API tests stay in
  integration.rs; their imports move to the crate-root re-exports.
- #[allow(dead_code)] on items the retraction revealed as internal-only or
  intentionally-unwired (e.g. the parity-kept async UDP loops); triage in #125.

Breaking; lands while 0.7.0 is unpublished. Closes #67.
