+++ riperf3/tests/integration.rs
@@ -6,7 +6,7 @@
 use std::sync::atomic::{AtomicU16, Ordering};
 use std::time::Duration;
 
-use riperf3::protocol::TransportProtocol;
+use riperf3::TransportProtocol;
 use riperf3::{ClientBuilder, ServerBuilder};
 
 /// Allocate unique ports for parallel test execution.
@@ -818,7 +818,7 @@ async fn tcp_congestion_applied() {
 #[cfg(target_os = "linux")]
 #[tokio::test]
 async fn cpu_affinity_applied() {
-    use riperf3::net::set_cpu_affinity;
+    use riperf3::set_cpu_affinity;
 
     // Set affinity to CPU 0 on the current thread and verify
     set_cpu_affinity(0).unwrap();
@@ -831,449 +831,10 @@ async fn cpu_affinity_applied() {
     );
 }
 
-// ---------------------------------------------------------------------------
-// TestConfig from params
-// ---------------------------------------------------------------------------
-
-mod test_config_tests {
-    use riperf3::protocol::{TestParams, TransportProtocol};
-    use riperf3::TestConfig;
-
-    #[test]
-    fn tcp_defaults() {
-        let p = TestParams {
-            tcp: Some(true),
-            ..Default::default()
-        };
-        let cfg = TestConfig::from_params(&p);
-        assert_eq!(cfg.protocol, TransportProtocol::Tcp);
-        assert_eq!(cfg.duration, 10);
-        assert_eq!(cfg.num_streams, 1);
-        assert_eq!(cfg.blksize, 128 * 1024);
-        assert!(!cfg.reverse);
-        assert!(!cfg.bidir);
-    }
-
-    #[test]
-    fn udp_defaults() {
-        let p = TestParams {
-            udp: Some(true),
-            ..Default::default()
-        };
-        let cfg = TestConfig::from_params(&p);
-        assert_eq!(cfg.protocol, TransportProtocol::Udp);
-        assert_eq!(cfg.blksize, 1460);
-    }
-
-    #[test]
-    fn full_params() {
-        let p = TestParams {
-            tcp: Some(true),
-            time: Some(30),
-            parallel: Some(4),
-            len: Some(65536),
-            reverse: Some(true),
-            bidirectional: Some(true),
-            omit: Some(2),
-            nodelay: Some(true),
-            mss: Some(1400),
-            window: Some(524288),
-            bandwidth: Some(1_000_000_000),
-            tos: Some(0x10),
-            congestion: Some("bbr".to_string()),
-            udp_counters_64bit: Some(1),
-            ..Default::default()
-        };
-        let cfg = TestConfig::from_params(&p);
-        assert_eq!(cfg.duration, 30);
-        assert_eq!(cfg.num_streams, 4);
-        assert_eq!(cfg.blksize, 65536);
-        assert!(cfg.reverse);
-        assert!(cfg.bidir);
-        assert_eq!(cfg.omit, 2);
-        assert!(cfg.no_delay);
-        assert_eq!(cfg.mss, Some(1400));
-        assert_eq!(cfg.window, Some(524288));
-        assert_eq!(cfg.bandwidth, 1_000_000_000);
-        assert_eq!(cfg.tos, 0x10);
-        assert_eq!(cfg.congestion, Some("bbr".to_string()));
-        assert!(cfg.udp_counters_64bit);
-    }
-
-    // -- server mirrors the client's rate resolution (#17) --
-
-    #[test]
-    fn udp_absent_bandwidth_is_unlimited() {
-        // iperf3 omits the `bandwidth` param only for -b 0 (unlimited) and sends
-        // it explicitly otherwise (incl. its 1 Mbit/s default); riperf3 clients
-        // always send it. So an ABSENT bandwidth means unlimited (0), matching
-        // iperf3 — NOT the 1 Mbit/s default. Defaulting to 1M throttled an
-        // iperf3 `-b 0` reverse/bidir client's server-side sender (#21).
-        let p = TestParams {
-            udp: Some(true),
-            ..Default::default()
-        };
-        let cfg = TestConfig::from_params(&p);
-        assert_eq!(cfg.bandwidth, 0);
-    }
-
-    #[test]
-    fn udp_explicit_zero_bandwidth_is_unlimited() {
-        // -b 0 carried in params → unlimited server-side, so reverse/bidir
-        // runs flat-out instead of being throttled to 1 Mbit/s.
-        let p = TestParams {
-            udp: Some(true),
-            bandwidth: Some(0),
-            ..Default::default()
-        };
-        let cfg = TestConfig::from_params(&p);
-        assert_eq!(cfg.bandwidth, 0);
-    }
-
-    #[test]
-    fn tcp_absent_bandwidth_is_unlimited() {
-        let p = TestParams {
-            tcp: Some(true),
-            ..Default::default()
-        };
-        let cfg = TestConfig::from_params(&p);
-        assert_eq!(cfg.bandwidth, 0);
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Protocol param/results round-trip
-// ---------------------------------------------------------------------------
-
-mod protocol_tests {
-    use riperf3::protocol::{self, StreamResultJson, TestParams, TestResultsJson};
-
-    #[tokio::test]
-    async fn params_round_trip() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let params = TestParams {
-            tcp: Some(true),
-            time: Some(30),
-            parallel: Some(4),
-            len: Some(65536),
-            reverse: Some(true),
-            nodelay: Some(true),
-            mss: Some(1400),
-            bandwidth: Some(1_000_000),
-            congestion: Some("cubic".to_string()),
-            client_version: Some("test 1.0".to_string()),
-            ..Default::default()
-        };
-
-        let params_clone = params.clone();
-        let writer = tokio::spawn(async move {
-            let mut stream = tokio::net::TcpStream::connect(addr).await.unwrap();
-            protocol::send_params(&mut stream, &params_clone)
-                .await
-                .unwrap();
-        });
-
-        let (mut stream, _) = listener.accept().await.unwrap();
-        let received = protocol::recv_params(&mut stream).await.unwrap();
-        writer.await.unwrap();
-
-        assert_eq!(received.tcp, Some(true));
-        assert_eq!(received.time, Some(30));
-        assert_eq!(received.parallel, Some(4));
-        assert_eq!(received.len, Some(65536));
-        assert_eq!(received.reverse, Some(true));
-        assert_eq!(received.nodelay, Some(true));
-        assert_eq!(received.mss, Some(1400));
-        assert_eq!(received.bandwidth, Some(1_000_000));
-        assert_eq!(received.congestion, Some("cubic".to_string()));
-        assert_eq!(received.client_version, Some("test 1.0".to_string()));
-        // Unset fields should remain None
-        assert_eq!(received.udp, None);
-        assert_eq!(received.bidirectional, None);
-    }
-
-    #[tokio::test]
-    async fn results_round_trip() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let results = TestResultsJson {
-            cpu_util_total: 42.5,
-            cpu_util_user: 30.0,
-            cpu_util_system: 12.5,
-            sender_has_retransmits: 3,
-            congestion_used: Some("bbr".to_string()),
-            streams: vec![
-                StreamResultJson {
-                    id: 1,
-                    bytes: 10_000_000,
-                    retransmits: 3,
-                    jitter: 0.0,
-                    errors: 0,
-                    omitted_errors: 0,
-                    packets: 0,
-                    omitted_packets: 0,
-                    start_time: 0.0,
-                    end_time: 10.0,
-                },
-                StreamResultJson {
-                    id: 3,
-                    bytes: 9_500_000,
-                    retransmits: 0,
-                    jitter: 0.0,
-                    errors: 0,
-                    omitted_errors: 0,
-                    packets: 0,
-                    omitted_packets: 0,
-                    start_time: 0.0,
-                    end_time: 10.0,
-                },
-            ],
-        };
-
-        let results_clone = results.clone();
-        let writer = tokio::spawn(async move {
-            let mut stream = tokio::net::TcpStream::connect(addr).await.unwrap();
-            protocol::send_results(&mut stream, &results_clone)
-                .await
-                .unwrap();
-        });
-
-        let (mut stream, _) = listener.accept().await.unwrap();
-        let received = protocol::recv_results(&mut stream).await.unwrap();
-        writer.await.unwrap();
-
-        assert_eq!(received.cpu_util_total, 42.5);
-        assert_eq!(received.sender_has_retransmits, 3);
-        assert_eq!(received.congestion_used, Some("bbr".to_string()));
-        assert_eq!(received.streams.len(), 2);
-        assert_eq!(received.streams[0].id, 1);
-        assert_eq!(received.streams[0].bytes, 10_000_000);
-        assert_eq!(received.streams[1].id, 3);
-    }
-
-    #[tokio::test]
-    async fn oversized_json_rejected() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let writer = tokio::spawn(async move {
-            let mut stream = tokio::net::TcpStream::connect(addr).await.unwrap();
-            // Write a JSON payload larger than MAX_PARAMS_JSON_LEN (8KB)
-            let big = serde_json::json!({"data": "x".repeat(10_000)});
-            protocol::json_write(&mut stream, &big).await.unwrap();
-        });
-
-        let (mut stream, _) = listener.accept().await.unwrap();
-        let result = protocol::json_read(&mut stream, protocol::MAX_PARAMS_JSON_LEN).await;
-        writer.await.unwrap();
-
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn params_with_unknown_fields_accepted() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        // Simulate a future iperf3 version sending extra fields
-        let writer = tokio::spawn(async move {
-            let mut stream = tokio::net::TcpStream::connect(addr).await.unwrap();
-            let json = serde_json::json!({
-                "tcp": true,
-                "time": 10,
-                "some_future_field": 42,
-                "another_new_thing": "hello"
-            });
-            protocol::json_write(&mut stream, &json).await.unwrap();
-        });
-
-        let (mut stream, _) = listener.accept().await.unwrap();
-        let result = protocol::recv_params(&mut stream).await;
-        writer.await.unwrap();
-
-        // Should succeed — unknown fields are ignored by serde default
-        let params = result.unwrap();
-        assert_eq!(params.tcp, Some(true));
-        assert_eq!(params.time, Some(10));
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Error path tests
-// ---------------------------------------------------------------------------
-
-mod error_tests {
-    use riperf3::utils::{parse_bitrate, parse_kmg};
-    use riperf3::ClientBuilder;
-    use riperf3::ConfigError;
-
-    #[test]
-    fn build_without_host_fails() {
-        let r = ClientBuilder::default().build();
-        assert!(matches!(r, Err(ConfigError::MissingField("host"))));
-    }
-
-    #[test]
-    fn parse_kmg_negative() {
-        assert!(parse_kmg("-1").is_err());
-    }
-
-    #[test]
-    fn parse_kmg_fractional() {
-        // #73: fractional suffixes are now accepted (iperf3 parses with %lf).
-        // 1.5 * 1024^2 = 1_572_864.
-        assert_eq!(parse_kmg("1.5M").unwrap(), 1_572_864);
-        // Still rejects genuinely-malformed numbers.
-        assert!(parse_kmg("1.5.5M").is_err());
-    }
-
-    #[test]
-    fn parse_kmg_empty_suffix() {
-        assert!(parse_kmg("K").is_err());
-    }
-
-    #[test]
-    fn parse_bitrate_empty() {
-        assert!(parse_bitrate("").is_err());
-    }
-
-    #[test]
-    fn parse_bitrate_bad_burst() {
-        assert!(parse_bitrate("100M/abc").is_err());
-    }
-
-    // -- Error type display messages --
-
-    #[test]
-    fn error_display_variants() {
-        use riperf3::RiperfError;
-        assert_eq!(
-            format!("{}", RiperfError::CookieMismatch),
-            "cookie mismatch"
-        );
-        assert_eq!(
-            format!("{}", RiperfError::AccessDenied),
-            "access denied by server"
-        );
-        assert_eq!(
-            format!("{}", RiperfError::PeerDisconnected),
-            "peer disconnected"
-        );
-        assert!(format!("{}", RiperfError::Aborted("test".into())).contains("test"));
-        assert_eq!(
-            format!("{}", RiperfError::ConnectionTimeout),
-            "connection timed out"
-        );
-        assert!(format!("{}", RiperfError::Protocol("bad".into())).contains("bad"));
-        assert!(format!("{}", RiperfError::Aborted("reason".into())).contains("reason"));
-    }
-
-    // -- Edge cases --
-
-    #[tokio::test]
-    async fn connect_to_wrong_port_fails() {
-        let client = ClientBuilder::new("127.0.0.1")
-            .port(Some(1)) // port 1 — almost certainly not listening
-            .duration(1)
-            .connect_timeout(std::time::Duration::from_millis(500))
-            .build()
-            .unwrap();
-        let result = client.run().await;
-        assert!(result.is_err(), "connecting to port 1 should fail");
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Protocol error state tests
-// ---------------------------------------------------------------------------
-
-mod protocol_error_tests {
-    use riperf3::protocol::{self, TestState};
-
-    #[tokio::test]
-    async fn client_handles_access_denied() {
-        // Server sends AccessDenied state — client should return AccessDenied error
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let server_task = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            // Read cookie (37 bytes)
-            let mut cookie = [0u8; 37];
-            tokio::io::AsyncReadExt::read_exact(&mut stream, &mut cookie)
-                .await
-                .unwrap();
-            // Send AccessDenied
-            protocol::send_state(&mut stream, TestState::AccessDenied)
-                .await
-                .unwrap();
-        });
-
-        let client = riperf3::ClientBuilder::new("127.0.0.1")
-            .port(Some(addr.port()))
-            .duration(1)
-            .build()
-            .unwrap();
-        let result = client.run().await;
-        assert!(result.is_err(), "client should error on AccessDenied");
-        let err = format!("{}", result.unwrap_err());
-        assert!(
-            err.contains("access denied") || err.contains("protocol"),
-            "error should mention access denied, got: {err}"
-        );
-        let _ = server_task.await;
-    }
-
-    #[tokio::test]
-    async fn client_handles_server_error() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let server_task = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            let mut cookie = [0u8; 37];
-            tokio::io::AsyncReadExt::read_exact(&mut stream, &mut cookie)
-                .await
-                .unwrap();
-            protocol::send_state(&mut stream, TestState::ServerError)
-                .await
-                .unwrap();
-        });
-
-        let client = riperf3::ClientBuilder::new("127.0.0.1")
-            .port(Some(addr.port()))
-            .duration(1)
-            .build()
-            .unwrap();
-        let result = client.run().await;
-        assert!(result.is_err(), "client should error on ServerError");
-        let _ = server_task.await;
-    }
-
-    #[tokio::test]
-    async fn client_handles_peer_disconnect_during_handshake() {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let server_task = tokio::spawn(async move {
-            let (stream, _) = listener.accept().await.unwrap();
-            // Accept connection then immediately close it
-            drop(stream);
-        });
-
-        let client = riperf3::ClientBuilder::new("127.0.0.1")
-            .port(Some(addr.port()))
-            .duration(1)
-            .build()
-            .unwrap();
-        let result = client.run().await;
-        assert!(result.is_err(), "client should error on peer disconnect");
-        let _ = server_task.await;
-    }
-}
+// test_config_tests migrated in-crate to src/server.rs (#67).
+// protocol_tests migrated in-crate to src/protocol.rs (#67).
+// error_tests migrated in-crate to src/utils.rs (#67).
+// protocol_error_tests migrated in-crate to src/protocol.rs (#67).
 
 // ---------------------------------------------------------------------------
 // JSON output validation
@@ -1307,45 +868,11 @@ mod json_output_tests {
         let _ = server_task.await;
     }
 
-    #[test]
-    fn test_params_serializes_all_fields() {
-        use riperf3::protocol::TestParams;
-        let p = TestParams {
-            tcp: Some(true),
-            time: Some(10),
-            parallel: Some(4),
-            len: Some(131072),
-            reverse: Some(true),
-            bidirectional: Some(true),
-            nodelay: Some(true),
-            mss: Some(1400),
-            window: Some(524288),
-            bandwidth: Some(1_000_000),
-            tos: Some(16),
-            congestion: Some("bbr".to_string()),
-            client_version: Some("riperf3 0.1.0".to_string()),
-            ..Default::default()
-        };
-        let json = serde_json::to_string(&p).unwrap();
-        let v: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(v["tcp"], true);
-        assert_eq!(v["time"], 10);
-        assert_eq!(v["parallel"], 4);
-        assert_eq!(v["len"], 131072);
-        assert_eq!(v["reverse"], true);
-        assert_eq!(v["bidirectional"], true);
-        assert_eq!(v["nodelay"], true);
-        assert_eq!(v["MSS"], 1400);
-        assert_eq!(v["window"], 524288);
-        assert_eq!(v["bandwidth"], 1_000_000);
-        assert_eq!(v["TOS"], 16);
-        assert_eq!(v["congestion"], "bbr");
-        assert_eq!(v["client_version"], "riperf3 0.1.0");
-    }
+    // test_params_serializes_all_fields migrated in-crate to src/protocol.rs (#67).
 
     #[test]
     fn test_results_json_structure() {
-        use riperf3::protocol::{StreamResultJson, TestResultsJson};
+        use riperf3::{StreamResultJson, TestResultsJson};
         let r = TestResultsJson {
             cpu_util_total: 50.0,
             cpu_util_user: 40.0,
@@ -1376,418 +903,11 @@ mod json_output_tests {
     }
 }
 
-// ---------------------------------------------------------------------------
-// Interval reporter edge cases
-// ---------------------------------------------------------------------------
+// interval_reporter_tests migrated in-crate to src/reporter.rs (#67).
 
-mod interval_reporter_tests {
-    use riperf3::protocol::TransportProtocol;
-    use riperf3::reporter::{spawn_interval_reporter, IntervalReporterConfig, ReporterEnd};
-    use std::sync::atomic::{AtomicBool, Ordering};
-    use std::sync::Arc;
+// udp_edge_tests migrated in-crate to src/stream.rs (#67).
 
-    #[tokio::test]
-    async fn disabled_returns_none() {
-        let done = Arc::new(AtomicBool::new(false));
-        let config = IntervalReporterConfig {
-            interval_secs: 0.0,
-            protocol: TransportProtocol::Tcp,
-            format_char: 'a',
-            omit_secs: 0,
-            num_streams: 1,
-            forceflush: false,
-            timestamp_format: None,
-            json_stream: false,
-            print: true,
-            blksize: 128 * 1024,
-        };
-        assert!(
-            spawn_interval_reporter(config, vec![], done, Arc::new(ReporterEnd::new()), None)
-                .is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn negative_interval_returns_none() {
-        let done = Arc::new(AtomicBool::new(false));
-        let config = IntervalReporterConfig {
-            interval_secs: -1.0,
-            protocol: TransportProtocol::Tcp,
-            format_char: 'a',
-            omit_secs: 0,
-            num_streams: 1,
-            forceflush: false,
-            timestamp_format: None,
-            json_stream: false,
-            print: true,
-            blksize: 128 * 1024,
-        };
-        assert!(
-            spawn_interval_reporter(config, vec![], done, Arc::new(ReporterEnd::new()), None)
-                .is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn zero_streams_doesnt_panic() {
-        let done = Arc::new(AtomicBool::new(false));
-        let config = IntervalReporterConfig {
-            interval_secs: 0.5,
-            protocol: TransportProtocol::Tcp,
-            format_char: 'a',
-            omit_secs: 0,
-            num_streams: 0,
-            forceflush: false,
-            timestamp_format: None,
-            json_stream: false,
-            print: true,
-            blksize: 128 * 1024,
-        };
-        let handle = spawn_interval_reporter(
-            config,
-            vec![],
-            done.clone(),
-            Arc::new(ReporterEnd::new()),
-            None,
-        );
-        assert!(handle.is_some());
-        // Let it tick once then stop via the `done` fallback (no `finish`).
-        tokio::time::sleep(std::time::Duration::from_millis(600)).await;
-        done.store(true, Ordering::Relaxed);
-        if let Some(h) = handle {
-            let _ = h.await;
-        }
-    }
-
-    /// #55: a test that ends part-way through an interval must still emit a
-    /// final partial interval carrying the residual bytes, rather than dropping
-    /// it. Drives two full intervals, then a partial third, and asserts the last
-    /// collected interval is the short partial (residual bytes, sub-interval
-    /// duration) — not a full interval and not missing entirely.
-    #[tokio::test]
-    async fn final_partial_interval_is_emitted() {
-        use riperf3::reporter::{CollectedIntervals, IntervalStreamRef};
-        use riperf3::stream::StreamCounters;
-        use std::sync::Mutex;
-        use std::time::Duration;
-
-        let interval = 0.5_f64;
-        let counters = Arc::new(StreamCounters::new());
-        let done = Arc::new(AtomicBool::new(false));
-        let collector = Arc::new(Mutex::new(CollectedIntervals::default()));
-
-        let stream_ref = IntervalStreamRef {
-            id: 1,
-            is_sender: true,
-            counters: counters.clone(),
-            udp_recv_stats: None,
-            raw_fd: None,
-        };
-        let config = IntervalReporterConfig {
-            interval_secs: interval,
-            protocol: TransportProtocol::Tcp,
-            format_char: 'a',
-            omit_secs: 0,
-            num_streams: 1,
-            forceflush: false,
-            timestamp_format: None,
-            json_stream: false,
-            print: false, // collect-only; assert on the collector
-            blksize: 128 * 1024,
-        };
-        let reporter_end = Arc::new(ReporterEnd::new());
-        let report_start = std::time::Instant::now();
-        let handle = spawn_interval_reporter(
-            config,
-            vec![stream_ref],
-            done.clone(),
-            reporter_end.clone(),
-            Some(collector.clone()),
-        )
-        .expect("reporter spawns for a positive interval");
-
-        // Two full intervals, each draining 1000 bytes at its tick. The 650ms
-        // sleeps give ~150ms slack so each tick lands before the next batch.
-        counters.record_sent(1000);
-        tokio::time::sleep(Duration::from_millis(650)).await; // tick @0.5 -> [0,0.5]=1000
-        counters.record_sent(1000);
-        tokio::time::sleep(Duration::from_millis(650)).await; // tick @1.0 -> [0.5,1.0]=1000
-
-        // Partial third interval: residual bytes, then end mid-interval (well
-        // before the @1.5 tick) by signalling the authoritative end time, exactly
-        // as the client/server driver does.
-        counters.record_sent(500);
-        tokio::time::sleep(Duration::from_millis(120)).await; // ~1.42s
-        reporter_end.finish(report_start.elapsed().as_secs_f64());
-        let _ = handle.await;
-
-        let g = collector.lock().unwrap();
-        let n = g.intervals.len();
-        assert!(n >= 1, "expected at least one collected interval");
-        let last = &g.intervals[n - 1];
-        // The defining property of the fix: the last interval is the short
-        // partial holding the 500 residual bytes — pre-fix it was dropped, so
-        // the last interval would be a full [0.5,1.0]=1000 instead.
-        assert_eq!(
-            last.sum.bytes, 500,
-            "final partial must carry the residual 500 bytes; got {} (intervals={n})",
-            last.sum.bytes
-        );
-        let dur = last.sum.end - last.sum.start;
-        assert!(
-            dur > 0.0 && dur < interval,
-            "final interval must be a sub-interval partial; start={} end={} dur={dur}",
-            last.sum.start,
-            last.sum.end
-        );
-    }
-
-    /// #55 guard: a run that ends exactly on an interval boundary must NOT emit a
-    /// trailing partial — even when the sender is still writing into the socket
-    /// at that instant (the saturating-TCP case). Records "slack" bytes after the
-    /// last tick to model the live sender, then finishes on the boundary; the
-    /// zero-length remainder must be dropped, so the slack bytes never surface as
-    /// a spurious interval.
-    #[tokio::test]
-    async fn no_spurious_partial_when_ending_on_boundary() {
-        use riperf3::reporter::{CollectedIntervals, IntervalStreamRef};
-        use riperf3::stream::StreamCounters;
-        use std::sync::Mutex;
-        use std::time::Duration;
-
-        let interval = 0.5_f64;
-        let counters = Arc::new(StreamCounters::new());
-        let done = Arc::new(AtomicBool::new(false));
-        let collector = Arc::new(Mutex::new(CollectedIntervals::default()));
-
-        let stream_ref = IntervalStreamRef {
-            id: 1,
-            is_sender: true,
-            counters: counters.clone(),
-            udp_recv_stats: None,
-            raw_fd: None,
-        };
-        let config = IntervalReporterConfig {
-            interval_secs: interval,
-            protocol: TransportProtocol::Tcp,
-            format_char: 'a',
-            omit_secs: 0,
-            num_streams: 1,
-            forceflush: false,
-            timestamp_format: None,
-            json_stream: false,
-            print: false,
-            blksize: 128 * 1024,
-        };
-        let reporter_end = Arc::new(ReporterEnd::new());
-        let handle = spawn_interval_reporter(
-            config,
-            vec![stream_ref],
-            done.clone(),
-            reporter_end.clone(),
-            Some(collector.clone()),
-        )
-        .expect("reporter spawns for a positive interval");
-
-        // Two full intervals.
-        counters.record_sent(1000);
-        tokio::time::sleep(Duration::from_millis(650)).await; // tick @0.5
-        counters.record_sent(1000);
-        tokio::time::sleep(Duration::from_millis(650)).await; // tick @1.0, now ~1.3s
-
-        // Live sender: bytes are still landing right as the run ends on the 1.0s
-        // boundary. The authoritative end time is exactly 1.0, so the remainder is
-        // zero-length and must be dropped despite these residual bytes.
-        counters.record_sent(777);
-        reporter_end.finish(1.0);
-        let _ = handle.await;
-
-        let g = collector.lock().unwrap();
-        assert_eq!(
-            g.intervals.len(),
-            2,
-            "no trailing partial when ending on a boundary; got {} intervals",
-            g.intervals.len()
-        );
-        let last = g.intervals.last().unwrap();
-        assert_eq!(
-            last.sum.bytes, 1000,
-            "last interval is the full [0.5,1.0]; the 777 slack bytes are not a new interval"
-        );
-    }
-
-    /// #55 receiver-side guard: a receiver whose `end_secs` trails the last
-    /// boundary (the control round-trip that delivers TEST_END) but whose tail
-    /// has no residual bytes (the peer already stopped) must not emit a trailing
-    /// empty interval. Mirrors the server's situation.
-    #[tokio::test]
-    async fn no_partial_for_receiver_with_no_residual() {
-        use riperf3::reporter::{CollectedIntervals, IntervalStreamRef};
-        use riperf3::stream::StreamCounters;
-        use std::sync::Mutex;
-        use std::time::Duration;
-
-        let interval = 0.5_f64;
-        let counters = Arc::new(StreamCounters::new());
-        let done = Arc::new(AtomicBool::new(false));
-        let collector = Arc::new(Mutex::new(CollectedIntervals::default()));
-
-        let stream_ref = IntervalStreamRef {
-            id: 1,
-            is_sender: false, // receiver side
-            counters: counters.clone(),
-            udp_recv_stats: None,
-            raw_fd: None,
-        };
-        let config = IntervalReporterConfig {
-            interval_secs: interval,
-            protocol: TransportProtocol::Tcp,
-            format_char: 'a',
-            omit_secs: 0,
-            num_streams: 1,
-            forceflush: false,
-            timestamp_format: None,
-            json_stream: false,
-            print: false,
-            blksize: 128 * 1024,
-        };
-        let reporter_end = Arc::new(ReporterEnd::new());
-        let handle = spawn_interval_reporter(
-            config,
-            vec![stream_ref],
-            done.clone(),
-            reporter_end.clone(),
-            Some(collector.clone()),
-        )
-        .expect("reporter spawns for a positive interval");
-
-        // One full interval of received bytes, then the peer stops: no further
-        // bytes arrive before the run ends.
-        counters.record_received(1000);
-        tokio::time::sleep(Duration::from_millis(650)).await; // tick @0.5 -> [0,0.5]=1000
-
-        // end_secs trails the 0.5 boundary (as TEST_END would), but the tail is
-        // empty — the residual-bytes guard must drop it.
-        reporter_end.finish(0.62);
-        let _ = handle.await;
-
-        let g = collector.lock().unwrap();
-        assert_eq!(
-            g.intervals.len(),
-            1,
-            "no trailing empty interval for an idle receiver tail; got {} intervals",
-            g.intervals.len()
-        );
-    }
-}
-
-// ---------------------------------------------------------------------------
-// UDP edge cases
-// ---------------------------------------------------------------------------
-
-mod udp_edge_tests {
-    use riperf3::stream::{UdpHeader, UdpRecvStats};
-
-    #[test]
-    fn udp_header_32bit_sequence_max() {
-        let h = UdpHeader {
-            sec: 0,
-            usec: 0,
-            seq: u32::MAX as u64,
-        };
-        let mut buf = [0u8; 16];
-        h.write_to(&mut buf, false);
-        let h2 = UdpHeader::read_from(&buf, false).unwrap();
-        assert_eq!(h2.seq, u32::MAX as u64);
-    }
-
-    #[test]
-    fn udp_header_64bit_sequence_max() {
-        let h = UdpHeader {
-            sec: 0,
-            usec: 0,
-            seq: u64::MAX,
-        };
-        let mut buf = [0u8; 16];
-        h.write_to(&mut buf, true);
-        let h2 = UdpHeader::read_from(&buf, true).unwrap();
-        assert_eq!(h2.seq, u64::MAX);
-    }
-
-    #[test]
-    fn udp_stats_massive_gap() {
-        // Simulate losing 1000 packets at once
-        let mut stats = UdpRecvStats::new();
-        stats.update(
-            &UdpHeader {
-                sec: 0,
-                usec: 0,
-                seq: 1,
-            },
-            0.0,
-        );
-        stats.update(
-            &UdpHeader {
-                sec: 0,
-                usec: 0,
-                seq: 1002,
-            },
-            1.0,
-        );
-        assert_eq!(stats.cnt_error, 1000);
-        assert_eq!(stats.packet_count, 1002);
-    }
-
-    #[test]
-    fn udp_stats_duplicate_packet() {
-        let mut stats = UdpRecvStats::new();
-        stats.update(
-            &UdpHeader {
-                sec: 0,
-                usec: 0,
-                seq: 1,
-            },
-            0.0,
-        );
-        stats.update(
-            &UdpHeader {
-                sec: 0,
-                usec: 0,
-                seq: 2,
-            },
-            0.001,
-        );
-        // Duplicate of packet 1
-        stats.update(
-            &UdpHeader {
-                sec: 0,
-                usec: 0,
-                seq: 1,
-            },
-            0.002,
-        );
-        assert_eq!(stats.outoforder_packets, 1);
-        assert_eq!(stats.packet_count, 2);
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Stream ID assignment
-// ---------------------------------------------------------------------------
-
-mod stream_id_tests {
-    use riperf3::utils::iperf3_stream_id;
-
-    #[test]
-    fn matches_iperf3_pattern() {
-        assert_eq!(iperf3_stream_id(0), 1);
-        assert_eq!(iperf3_stream_id(1), 3);
-        assert_eq!(iperf3_stream_id(2), 4);
-        assert_eq!(iperf3_stream_id(3), 5);
-        assert_eq!(iperf3_stream_id(4), 6);
-        assert_eq!(iperf3_stream_id(9), 11);
-    }
-}
+// stream_id_tests migrated in-crate to src/utils.rs (#67).
 
 // ===========================================================================
 // Unimplemented flag tests
@@ -1835,17 +955,7 @@ mod implemented_flag_tests {
         let _ = server_task.await;
     }
 
-    #[test]
-    fn repeating_payload_buffer() {
-        let buf = riperf3::utils::make_send_buffer(256, true);
-        assert_eq!(buf.len(), 256);
-        assert_eq!(buf[0], 0);
-        assert_eq!(buf[1], 1);
-        assert_eq!(buf[255], 255);
-
-        let zeros = riperf3::utils::make_send_buffer(256, false);
-        assert!(zeros.iter().all(|&b| b == 0));
-    }
+    // repeating_payload_buffer migrated in-crate to src/utils.rs (#67).
 
     #[tokio::test]
     async fn repeating_payload_runs() {
@@ -1958,7 +1068,7 @@ mod implemented_flag_tests {
     #[test]
     fn cpu_affinity_readback() {
         // Verify set_cpu_affinity actually works via getaffinity readback
-        use riperf3::net::set_cpu_affinity;
+        use riperf3::set_cpu_affinity;
         set_cpu_affinity(0).unwrap();
         let cpuset = nix::sched::sched_getaffinity(nix::unistd::Pid::from_raw(0)).unwrap();
         assert!(cpuset.is_set(0).unwrap());
@@ -2048,21 +1158,7 @@ mod implemented_flag_tests {
         let _ = server_task.await;
     }
 
-    #[test]
-    fn dscp_symbolic_and_numeric() {
-        use riperf3::utils::parse_dscp;
-        // Symbolic names
-        assert_eq!(parse_dscp("ef").unwrap(), 46 << 2); // EF = 184
-        assert_eq!(parse_dscp("af11").unwrap(), 10 << 2); // AF11 = 40
-        assert_eq!(parse_dscp("cs1").unwrap(), 8 << 2); // CS1 = 32
-                                                        // Numeric
-        assert_eq!(parse_dscp("46").unwrap(), 46 << 2);
-        assert_eq!(parse_dscp("0x2e").unwrap(), 46 << 2); // 0x2e = 46
-        assert_eq!(parse_dscp("056").unwrap(), 46 << 2); // 056 octal = 46
-                                                         // Out of range
-        assert!(parse_dscp("64").is_err());
-        assert!(parse_dscp("abc").is_err());
-    }
+    // dscp_symbolic_and_numeric migrated in-crate to src/utils.rs (#67).
 
     #[tokio::test]
     async fn dscp_flag_runs() {
@@ -3144,7 +2240,6 @@ async fn udp_sendmmsg_with_64bit_counters() {
 
 mod client_run_return_value {
     use super::*;
-    use riperf3::RiperfError;
 
     /// Happy path: a normal TCP exchange yields a populated `TestResultsJson`,
     /// proving the library now exposes what `print_results` used to consume internally.
@@ -3184,47 +2279,5 @@ mod client_run_return_value {
 
         let _ = server_task.await;
     }
-
-    /// Error path: a server that ends the session via `IperfDone` without an
-    /// `ExchangeResults` round now yields `Protocol("missing server results...")`
-    /// instead of the previous `Ok(())`. Uses a mock TCP server because the real
-    /// riperf3 server always performs `ExchangeResults`.
-    #[tokio::test]
-    async fn run_errors_when_server_skips_results_exchange() {
-        use riperf3::protocol::{self, TestState};
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        let server_task = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            let mut cookie = [0u8; 37];
-            tokio::io::AsyncReadExt::read_exact(&mut stream, &mut cookie)
-                .await
-                .unwrap();
-            // Skip ExchangeResults / DisplayResults entirely.
-            protocol::send_state(&mut stream, TestState::IperfDone)
-                .await
-                .unwrap();
-        });
-
-        let client = ClientBuilder::new("127.0.0.1")
-            .port(Some(addr.port()))
-            .duration(1)
-            .build()
-            .unwrap();
-        let err = client
-            .run()
-            .await
-            .expect_err("expected missing-results error");
-        match err {
-            RiperfError::Protocol(msg) => assert!(
-                msg.contains("missing server results"),
-                "unexpected protocol message: {msg}"
-            ),
-            other => panic!("expected RiperfError::Protocol, got {other:?}"),
-        }
-
-        let _ = server_task.await;
-    }
+    // run_errors_when_server_skips_results_exchange migrated in-crate to src/client.rs (#67).
 }
